ABS.Compiler.CmdOpt

Description

Possible options passed to The ABS-Haskell compiler

Documentation

cmdOpt :: CmdOpt

data CmdOpt

Constructors

CmdOpt

Fields

abs_sources :: [FilePath]

The input ABS module files (ending in .abs) or ABS directories

output_dir :: FilePath

In which directory to put all the Haskell translated files (.hs files)

create_script :: Bool

creates a helper bash script for easier invoking the ghc Haskell compiler and puts it in the outputdir

dump_ast :: Bool

A flag to dump the parsed AST to stderr

nostdlib :: Bool

cmdOptSpec :: CmdOpt