ABS.Compiler.Codegen.StmExp

Description

abs-pure-expressions in the statement world (which allows mutable local variables, fields, etc)

the resulting haskell expressions have type IO, which means that eventually later have to be lifted to ABS monad

Documentation

tStmExp :: (?st :: SymbolTable, ?vars :: ScopeLVL, ?fields :: ScopeLVL, ?cname :: String) => PureExp -> LetScope (Exp, T)

Translating a pure expression augmented to work with mutable local variables & fields of the statement world

mUpOne :: (?st :: SymbolTable) => T -> T -> Exp -> Exp

mUpMany :: (?st :: SymbolTable) => [T] -> [T] -> [Exp] -> [Exp]