ABS.Compiler.Firstpass.Base

Documentation

type ModuleName = String

type SymbolTables = Map ModuleName SymbolTable

A mapping of Modules to their SymbolTables

type SymbolTable = Map SymbolName SymbolValue

A separate SymbolTable for each module

data SymbolName

Name of the Symbol, coupled with its origin, and if it is qualified-imported

Constructors

SN String (Maybe (ModuleName, IsQualified))

data SymbolValue

The entry of the Symbol, its symbol type and if it is exported

Constructors

SV SymbolType IsExported

data SymbolType

The different symbol types

Constructors

Function [U] [T] T
tyvars, input types, output type

Datatype

Datacons String [U] [T] T
from which datatype it comes (required by Haskell module system), tyvars, input types, output type

Exception

Class [T] [T]

Interface [(String, Maybe [String])] (Map SymbolName [(String, Maybe [String])])
its direct method names, http-callable formal parameters & map of *all* extends interfaces to their own methods, http-callable formal parameters

Foreign

type IsQualified = Bool

type IsExported = Bool